Refactor service request update navigation by removing HasServiceRequestSubNavigation trait and implementing sub-navigation directly in resource class
