--- app-modules/service-management/src/Filament/Concerns/HasServiceRequestSubNavigation.php
@@ -1,78 +1,0 @@
-<?php
-
-/*
-<COPYRIGHT>
-
-    Copyright © 2016-2026, Canyon GBS Inc. All rights reserved.
-
-    Aiding App® is licensed under the Elastic License 2.0. For more details,
-    see <https://github.com/canyongbs/aidingapp/blob/main/LICENSE.>
-
-    Notice:
-
-    - You may not provide the software to third parties as a hosted or managed
-      service, where the service provides users with access to any substantial set of
-      the features or functionality of the software.
-    - You may not move, change, disable, or circumvent the license key functionality
-      in the software, and you may not remove or obscure any functionality in the
-      software that is protected by the license key.
-    - You may not alter, remove, or obscure any licensing, copyright, or other notices
-      of the licensor in the software. Any use of the licensor's trademarks is subject
-      to applicable law.
-    - Canyon GBS Inc. respects the intellectual property rights of others and expects the
-      same in return. Canyon GBS® and Aiding App® are registered trademarks of
-      Canyon GBS Inc., and we are committed to enforcing and protecting our trademarks
-      vigorously.
-    - The software solution, including services, infrastructure, and code, is offered as a
-      Software as a Service (SaaS) by Canyon GBS Inc.
-    - Use of this software implies agreement to the license terms and conditions as stated
-      in the Elastic License 2.0.
-
-    For more information or inquiries please visit our website at
-    <https://www.canyongbs.com> or contact us via email at legal@canyongbs.com.
-
-</COPYRIGHT>
-*/
-
-namespace AidingApp\ServiceManagement\Filament\Concerns;
-
-use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\EditServiceRequest;
-use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\Feedback;
-use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ManageAssignments;
-use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ManageServiceRequestFormSubmission;
-use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ManageServiceRequestUpdate;
-use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ServiceRequestTimeline;
-use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ViewServiceRequest;
-use App\Enums\Feature;
-use Illuminate\Support\Facades\Gate;
-
-trait HasServiceRequestSubNavigation
-{
-    public function getSubNavigation(): array
-    {
-        $serviceRequest = $this->getParentRecord();
-
-        $navigationItems = [
-            ViewServiceRequest::class,
-            EditServiceRequest::class,
-            ManageAssignments::class,
-            ManageServiceRequestUpdate::class,
-            ServiceRequestTimeline::class,
-        ];
-
-        if ($serviceRequest->serviceRequestFormSubmission) {
-            array_splice($navigationItems, 1, 0, ManageServiceRequestFormSubmission::class);
-        }
-
-        if (Gate::check(Feature::FeedbackManagement->getGateName())) {
-            array_splice($navigationItems, array_search(ServiceRequestTimeline::class, $navigationItems), 0, Feedback::class);
-        }
-
-        return $this->generateNavigationItems($navigationItems);
-    }
-
-    public function getSubNavigationParameters(): array
-    {
-        return ['record' => $this->getParentRecord()];
-    }
-}
