+++ app-modules/service-management/src/Filament/Resources/ServiceRequestUpdates/Pages/EditServiceRequestUpdate.php
@@ -36,7 +36,6 @@
 
 namespace AidingApp\ServiceManagement\Filament\Resources\ServiceRequestUpdates\Pages;
 
-use AidingApp\ServiceManagement\Filament\Concerns\HasServiceRequestSubNavigation;
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequestUpdates\ServiceRequestUpdateResource;
 use App\Concerns\EditPageRedirection;
 use Filament\Actions\DeleteAction;
@@ -46,9 +45,13 @@ use Filament\Resources\Pages\EditRecord;
 class EditServiceRequestUpdate extends EditRecord
 {
     use EditPageRedirection;
-    use HasServiceRequestSubNavigation;
 
     protected static string $resource = ServiceRequestUpdateResource::class;
+
+    public function getSubNavigationParameters(): array
+    {
+        return ['record' => $this->getParentRecord()];
+    }
 
     public function getRedirectUrl(): ?string
     {
