+++ app-modules/service-management/src/Filament/Resources/ServiceRequestUpdates/ServiceRequestUpdateResource.php
@@ -36,11 +36,19 @@
 
 namespace AidingApp\ServiceManagement\Filament\Resources\ServiceRequestUpdates;
 
+use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\EditServiceRequest;
+use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\Feedback;
+use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ManageAssignments;
+use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ManageServiceRequestFormSubmission;
+use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ManageServiceRequestUpdate;
+use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ServiceRequestTimeline;
+use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\Pages\ViewServiceRequest;
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequests\ServiceRequestResource;
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequestUpdates\Pages\EditServiceRequestUpdate;
 use AidingApp\ServiceManagement\Filament\Resources\ServiceRequestUpdates\Pages\ViewServiceRequestUpdate;
 use AidingApp\ServiceManagement\Models\ServiceRequest;
 use AidingApp\ServiceManagement\Models\ServiceRequestUpdate;
+use App\Enums\Feature;
 use App\Filament\Tables\Columns\IdColumn;
 use BackedEnum;
 use Filament\Actions\BulkActionGroup;
@@ -51,6 +59,7 @@ use Filament\Forms\Components\Select;
 use Filament\Forms\Components\SpatieMediaLibraryFileUpload;
 use Filament\Forms\Components\Textarea;
 use Filament\Forms\Components\Toggle;
+use Filament\Resources\Pages\Page;
 use Filament\Resources\Resource;
 use Filament\Schemas\Components\Section;
 use Filament\Schemas\Schema;
@@ -59,6 +68,7 @@ use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Filters\TernaryFilter;
 use Filament\Tables\Table;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Facades\Gate;
 
 class ServiceRequestUpdateResource extends Resource
 {
@@ -143,6 +153,31 @@ class ServiceRequestUpdateResource extends Resource
                     DeleteBulkAction::make(),
                 ]),
             ]);
+    }
+
+    public static function getRecordSubNavigation(Page $page): array
+    {
+        $serviceRequest = $page->getParentRecord();
+
+        assert($serviceRequest instanceof ServiceRequest);
+
+        $navigationItems = [
+            ViewServiceRequest::class,
+            EditServiceRequest::class,
+            ManageAssignments::class,
+            ManageServiceRequestUpdate::class,
+            ServiceRequestTimeline::class,
+        ];
+
+        if ($serviceRequest->serviceRequestFormSubmission) {
+            array_splice($navigationItems, 1, 0, ManageServiceRequestFormSubmission::class);
+        }
+
+        if (Gate::check(Feature::FeedbackManagement->getGateName())) {
+            array_splice($navigationItems, array_search(ServiceRequestTimeline::class, $navigationItems), 0, Feedback::class);
+        }
+
+        return $page->generateNavigationItems($navigationItems);
     }
 
     public static function getRelations(): array
